supp_app_drf_v2/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
drf-yasg==1.21.4
django_debug_toolbar==3.8.1
python-dotenv==0.21.0
celery==5.2.7
redis==4.3.4
python-dotenv==0.21.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
